# server/openapi.yaml
openapi: 3.0.1
info:
  title: HireReady API
  description: OpenAPI documentation for HireReady backend
  version: 1.0.0
servers:
  - url: http://localhost:5000/api
paths:
  /health:
    get:
      summary: Health check
      responses:
        '200':
          description: OK
  /questions:
    get:
      summary: List questions
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/Question'
    post:
      summary: Create question (admin)
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/Question'
      responses:
        '201':
          description: Created
  /interviews:
    post:
      summary: Create interview
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                category:
                  type: string
                difficulty:
                  type: string
      responses:
        '201':
          description: Interview created
  /auth/register:
    post:
      summary: Register new user
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                firstName:
                  type: string
                lastName:
                  type: string
                email:
                  type: string
                password:
                  type: string
      responses:
        '201':
          description: User created
  /auth/login:
    post:
      summary: Login user
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                email:
                  type: string
                password:
                  type: string
      responses:
        '200':
          description: Login successful
        '401':
          description: Invalid credentials
  /users/me:
    get:
      summary: Get current authenticated user
      security:
        - bearerAuth: []
      responses:
        '200':
          description: User object
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  schemas:
    User:
      type: object
      properties:
        id:
          type: string
        firstName:
          type: string
        lastName:
          type: string
        email:
          type: string
        role:
          type: string
    Question:
      type: object
      properties:
        id:
          type: string
        title:
          type: string
        body:
          type: string
        type:
          type: string
        difficulty:
          type: string
    Interview:
      type: object
      properties:
        id:
          type: string
        userId:
          type: string
        status:
          type: string
        totalScore:
          type: number